Move Objective-C compatibility layer back to TelemetryClient target
